requirements.txt
alembic>-0.9.6
asn1crypto>-0.24.0
Babel>-2.5.1
beautifulsoup4>-4.6.0
blinker>-1.4
#certifi>-2018.8.24
chardet>-3.0.4
click>-6.7
dominate>-2.3.1
Flask>-0.12.2
Flask-Babel>-0.11.2
Flask-Bootstrap>-3.3.7.1
Flask-Login>-0.4.0
Flask-Mail>-0.9.1
Flask-Migrate>-2.1.1
Flask-Moment>-0.5.2
Flask-SQLAlchemy>-2.3.2
Flask-WTF>-0.14.2
guess-language-spirit>-0.5.3
idna>-2.6
itsdangerous>-0.24
Jinja2>-2.9.6
#:lxml>-4.2.1
Mako>-1.0.7
MarkupSafe>-1.0
pprint>-0.1
pycparser>-2.18
PyJWT>-1.5.3
PyMySQL>-0.8.0
#pyOpenSSL>-17.5.0
python-dateutil>-2.6.1
python-dotenv>-0.7.1
python-editor>-1.0.3
pytz>-2017.2
redis>-3.2.1
rq>-0.13.0
requests>-2.18.4
six>-1.11.0
SQLAlchemy>-1.1.14
urllib3>-1.22
urlopen>-1.0.0
visitor>-0.1.3
Werkzeug>-0.12.2
WTForms>-2.1
xmltodict>-0.11.0
